AlteracoesNoCampoFuncionario

+++ src/SistemaWeb/Controllers/FuncionariosController.cs
@@ -126,7 +126,7 @@ namespace SistemaWeb.Controllers
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
-            {
+            {       
                 return NotFound();
             }
 
@@ -146,9 +146,19 @@ namespace SistemaWeb.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var funcionario = await _context.Funcionarios.FindAsync(id);
-            _context.Funcionarios.Remove(funcionario);
-            await _context.SaveChangesAsync();
+            var funcionario = await _context.Funcionarios.FindAsync(id);         
+            var agendamento = await _context.Agendamentos.FirstOrDefaultAsync(x => x.FuncionarioId.Equals(id));
+            if(agendamento == null)
+            {
+                _context.Funcionarios.Remove(funcionario);
+                await _context.SaveChangesAsync();
+            }
+            else
+            {
+                //return ViewBag = ("Teste");
+            }
+
+            
             return RedirectToAction(nameof(Index));
         }
